Fixed SQL for "exclude" parameter
IDs listed in the 'exclude' parameter were being put in a "NOT IN" clause, but not being removed from the "IN" clause. As a result those resources were not being properly excluded at least some of the time (not sure if this behavior was consistent across all environments).

I don't know if something similar may need to be done in simplesearchdriversolr.class.php.